requirements.txt
pydantic
httpx
beautifulsoup4
duckduckgo_search
google-genai
python-dotenv
